+++ src/components/applications/instances/InstanceService.php
@@ -33,7 +33,7 @@ class InstanceService extends Item implements IInstanceService
             Instance::FIELD__EVENTS => $app->getEvents(),
             Instance::FIELD__OPERATIONS => $app->getOperations(),
             Instance::FIELD__VENDOR => [IVendor::FIELD__NAME => $vendorName],
-            Instance::FIELD__CLASS => $app->buildOptions()->getResolver()
+            Instance::FIELD__CLASS => $app->getResolver()
         ]);
 
         /**